Moved state to PostsPage because user won't be able to see any page unless logged into App
